Add files via upload
Stage 2 finished, ready for playing friend against friend

Stage 2/4: The game is on_
Description
In this stage, we will add an event listener for our field cells. When you click on a free cell for the first time, add an X symbol to the cell. The second click on another free cell should add an O symbol. You will also need a status bar and a Reset button. A click on the Reset button should clear the field. These improvements can allow you to play Tic-tac-toe against your friend.

Objectives
You need to add a listener for each cell of the field. When you click on a free cell, the program should add X or O, depending on the move. If a user clicks on an occupied cell, ignore this click.

The program should track the status of the game. If the game is over, the program should ignore further clicks. At any time, when you press the Reset button, the program should clear the field and start a new game.

The status field should contain one of the following messages:

Game is not started
Game in progress
X wins
O wins
Draw
For testing purposes, add names for the following components:

LabelStatus for the component that displays the game status.
ButtonReset for the reset button.
The cell names should remain as in the previous stage. To display a symbol, set the button text. It must be one of the three strings: (one space) for an empty cell, X for the first player, and O for the second player.

To remove the border around the text of the active cell, use the setFocusPainted(false) method when creating a field cell component.

You can set any font, size, and color for cell symbols and any cell background. The tests only check the names of the components and their text.